+++ cmd/cilium-agent-proxy/app/helper.go
@@ -1,4 +1,4 @@
-package sub
+package app
 
 import (
 	"bytes"
