+++ cmd/cilium-agent-proxy/app/run.go
@@ -1,4 +1,4 @@
-package sub
+package app
 
 import (
 	"bytes"
