+++ cmd/cilium-agent-proxy/main.go
@@ -1,7 +1,7 @@
 package main
 
-import "github.com/cybozu-go/network-policy-viewer/cmd/cilium-agent-proxy/sub"
+import "github.com/cybozu-go/network-policy-viewer/cmd/cilium-agent-proxy/app"
 
 func main() {
-	sub.Execute()
+	app.Execute()
 }
