+++ cmd/npv/app/dump.go
@@ -1,4 +1,4 @@
-package sub
+package app
 
 import (
 	"bytes"
