+++ cmd/npv/app/list.go
@@ -1,4 +1,4 @@
-package sub
+package app
 
 import (
 	"context"
