+++ cmd/npv/app/root.go
@@ -1,4 +1,4 @@
-package sub
+package app
 
 import (
 	"fmt"
